put case objects in subdirectory
