+++ scripts/Tools/case.cmpgen_namelists
@@ -8,7 +8,6 @@ generate, or both) for this case.
 from standard_script_setup import *
 
 from CIME.case                  import Case
-from CIME.case_cmpgen_namelists import case_cmpgen_namelists
 
 ###############################################################################
 def parse_command_line(args, description):
@@ -60,7 +59,7 @@ def _main_func(description):
     caseroot, compare, generate, compare_name, generate_name, baseline_root \
         = parse_command_line(sys.argv, description)
     with Case(caseroot, read_only=True) as case:
-        success = case_cmpgen_namelists(case, compare, generate, compare_name, generate_name, baseline_root)
+        success = case.case_cmpgen_namelists(compare, generate, compare_name, generate_name, baseline_root)
 
     sys.exit(0 if success else CIME.utils.TESTS_FAILED_ERR_CODE)
