+++ scripts/Tools/check_case
@@ -8,8 +8,6 @@ from standard_script_setup import *
 
 from CIME.utils import expect
 from CIME.case  import Case
-from CIME.check_lockedfiles import check_lockedfiles
-from CIME.preview_namelists import create_namelists
 
 import argparse
 
@@ -44,8 +42,8 @@ def _main_func(description):
     parse_command_line(sys.argv, description)
 
     with Case(read_only=False) as case:
-        check_lockedfiles(case)
-        create_namelists(case)
+        case.check_lockedfiles()
+        case.create_namelists()
         build_complete = case.get_value("BUILD_COMPLETE")
 
     if not build_complete:
