+++ scripts/lib/CIME/SystemTests/irt.py
@@ -12,7 +12,6 @@ This test the model's restart capability as well as the short term archiver's in
 
 from CIME.SystemTests.restart_tests import RestartTest
 from CIME.XML.standard_module_setup import *
-from CIME.case_st_archive import case_st_archive, restore_from_archive
 from CIME.utils import ls_sorted_by_mtime
 
 logger = logging.getLogger(__name__)
@@ -28,13 +27,12 @@ class IRT(RestartTest):
                              multisubmit = False)
 
     def _case_one_custom_postrun_action(self):
-        case_st_archive(self._case)
+        self._case.case_st_archive()
         # Since preview namelist is run before _case_two_prerun_action, we need to do this here.
         dout_s_root = self._case1.get_value("DOUT_S_ROOT")
         restart_list = ls_sorted_by_mtime(os.path.join(dout_s_root,"rest"))
         logger.info("Restart directory list is {}".format(restart_list))
         expect(len(restart_list) >=2,"Expected at least two restart directories")
         # Get the older of the two restart directories
-        restore_from_archive(self._case2,
-                             rest_dir=os.path.abspath(
-                                 os.path.join(dout_s_root, "rest", restart_list[0])))
+        self._case2.restore_from_archive(rest_dir=os.path.abspath(
+            os.path.join(dout_s_root, "rest", restart_list[0])))
