+++ scripts/lib/CIME/case/case_cmpgen_namelists.py
