+++ scripts/lib/CIME/case/preview_namelists.py
